205--refactor and use window.API.onSet properly
